+++ example/test/widget_test.dart
@@ -1,30 +1,22 @@
-// This is a basic Flutter widget test.
-//
-// To perform an interaction with a widget in your test, use the WidgetTester
-// utility in the flutter_test package. For example, you can send tap and scroll
-// gestures. You can also use WidgetTester to find child widgets in the widget
-// tree, read text, and verify that the values of widget properties are correct.
-
-import 'package:flutter/material.dart';
 import 'package:flutter_test/flutter_test.dart';
-
 import 'package:flutter_svg_pro_example/main.dart';
 
 void main() {
-  testWidgets('Counter increments smoke test', (WidgetTester tester) async {
+  testWidgets('SVG Pro Diagnostics Smoke Test', (WidgetTester tester) async {
     // Build our app and trigger a frame.
     await tester.pumpWidget(const MyApp());
 
-    // Verify that our counter starts at 0.
-    expect(find.text('0'), findsOneWidget);
-    expect(find.text('1'), findsNothing);
+    // Verify that our app header shows the title.
+    expect(find.text('SVG PRO DIAGNOSTICS'), findsOneWidget);
 
-    // Tap the '+' icon and trigger a frame.
-    await tester.tap(find.byIcon(Icons.add));
-    await tester.pump();
+    // Verify that we display the list of views / sides.
+    expect(find.text('FRONT VIEW'), findsOneWidget);
+    expect(find.text('Left Side'), findsOneWidget);
+    expect(find.text('Top View'), findsOneWidget);
+    expect(find.text('Right Side'), findsOneWidget);
+    expect(find.text('Back View'), findsOneWidget);
 
-    // Verify that our counter has incremented.
-    expect(find.text('0'), findsNothing);
-    expect(find.text('1'), findsOneWidget);
+    // Verify that the initial selection count starts at 0.
+    expect(find.text('0 MARKED'), findsOneWidget);
   });
 }